make particle filter likelihood compatible with references or pointers
applications may require a mix of likelihood types (i.e., to model likelihoods for depth vs location data, or for missing data), which will require a likelihood sequence container with *pointers* to the likelihood implementations in order to avoid C++ object slicing issues when likelihoods are stored in STL containers.  so, it is helpful to allow the particle filter to accept a more flexible range of sequence types.